refactor: extract shared aiohttp session to src/utils/http.py
- New  with lazy  and  for
  a single aiohttp.ClientSession reused across the bot lifetime.
-  delegates session management to http utility.
-  switches from per-request sessions to the shared
  pool (connection reuse, less overhead).
-  shutdown cleanup updated to call .
